[Askbot] Refactor identity methods for Askbot backend and add anonymization.
This commit moves identity related methods from enriched/askbot
to identities/askbot. Support for identity anonymization
has also been added.
